Issue #1759 was partially fixed.

+++ core/src/main/java/com/orientechnologies/orient/core/index/hashindex/local/OLocalHashTable.java
@@ -1183,9 +1183,19 @@ public class OLocalHashTable<K, V> extends OSharedResourceAdaptive {
       final int index = bucket.getIndex(hashCode, key);
 
       if (index > -1) {
+        final int updateResult = bucket.updateEntry(index, value);
+        if (updateResult == 0)
+          return;
+
+        if (updateResult == 1) {
+          cacheEntry.markDirty();
+          return;
+        }
+
+        assert updateResult == -1;
+
         bucket.deleteEntry(index);
         size--;
-        cacheEntry.markDirty();
       }
 
       if (bucket.addEntry(hashCode, key, value)) {
